requirements.txt
pandas
numpy
scikit-learn
xgboost
streamlit
pyyaml
matplotlib
seaborn
joblib
kaggle
jupyter